+++ src/collective/dms/basecontent/indexers.py
@@ -1,11 +1,9 @@
-from ZODB.POSException import ConflictError
-from five import grok
+from collective.dms.basecontent.dmsdocument import IDmsDocument
 from OFS.interfaces import IItem
+from plone.indexer import indexer
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import base_hasattr
-from plone.indexer import indexer
-
-from .dmsdocument import IDmsDocument
+from ZODB.POSException import ConflictError
 
 
 @indexer(IDmsDocument)
@@ -39,9 +37,6 @@ def document_dynamic_searchable_text_indexer(obj):
                 raise
 
     return u' '.join(indexed_elements)
-
-grok.global_adapter(document_dynamic_searchable_text_indexer,
-                    name='SearchableText')
 
 
 @indexer(IItem)