+++ CHANGES.rst
@@ -6,7 +6,9 @@ Changelog
 
 - Removed `collective.z3cform.chosen` dependency.
   [sgeulette]
-- Removed grok in viewlets. Commented `dms.changetitle` viewlet definition.
+- Removed grok for viewlets. Commented `dms.changetitle` viewlet definition.
+  [sgeulette]
+- Removed grok for columns.
   [sgeulette]
 
 1.12 (2023-05-12)